requirements.txt
absl-py==1.0.0
ale-py==0.7.5
anylogic-alpyne==0.2.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
black==22.3.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1649586564473/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
gym==0.21.0
gym-notices==0.0.6
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel @ file:///D:/bld/ipykernel_1649684468721/work/dist/ipykernel-6.13.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1648413716235/work
jedi @ file:///D:/bld/jedi_1649067326950/work
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1650972844302/work
jupyter-core @ file:///D:/bld/jupyter_core_1645024505069/work
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mypy-extensions==0.4.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.9.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
protobuf==3.20.1
psutil @ file:///D:/bld/psutil_1648873144269/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
pywin32==303
pyzmq @ file:///D:/bld/pyzmq_1649055812949/work
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sb3-contrib @ git+https://github.com/Stable-Baselines-Team/stable-baselines3-contrib@db4c0114d07bb7439dc7ef02eef2a6e0d0de09bd
scipy==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stable-baselines==2.10.2
stable-baselines3==1.5.1a8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
torch==1.11.0+cu113
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tornado @ file:///D:/bld/tornado_1648827448708/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.1.2
zipp==3.8.0
